Added Desert Town Stuff
Created some things to do in the desert town.  Added shops to the map,
but haven't made the insides yet.  Introduced the party to a new
character, who I think turn out to be from the ice island on the world
map and potentially one of the Order of Five.