Lava Race / Anti Lag Script
The lava race has been completed, except that the secret reward is
currently only 1 Skal.  I personally think it should be a class ring.  I
found and added a Victor script which drastically reduces lag when there
are lots of parallel events.  Before the implementation of this script,
there were lag problems due to the large number of lava tiles.  The WTBA
document was updated.