Sprite Dump, Mostly figured out scripted change of actor graphic and face
Outside of battle I can successfully change an actor's character
graphic, face graphic (for the pause menu) and battle face graphic (for
the battle hud).

Inside battle I can change the character graphic and face graphic with
no problem.  I can also change the battle face graphic, but I'm having a
hard time getting the battle hud to refresh with the new image.

Basically, the battle graphic is loaded from a file called
Graphics/Huds/Battle/Face_<actorID>.png.  I was able to scrap together a
copy-paste function to change this file out successfully, and when I
swap the file before battle the graphic shows up correctly when the
battle starts.  But when the graphic is swapped DURING battle, the HUD
fails to refresh (even though in Windows explorer you can see that the
file has changed).  I'm trying to figure out a way to force the battle
HUD to refresh but running into trouble.

See the newest skill that Testy uses and the associated common event for
more details.